extract into revealanimation
